Create README.md
These R scripts outline the data cleaning and statistical modelling processes employed for the 2-stage Bayesian geostatistical bottom-up population modelling for Papua New Guniea. There are three main R scripts contained in the folder - covariates selection, model fitting and model cross-validation. The covariates selection script outlines the GLM-based stepwise regression approaches as well as the checks for multicollinearity we carried out to select the best set of geospatial covariates for the final model. The model fitting script shows the step-by-step approaches for the specification of the various models used in both the modelling of the building intensity and the population density. Finally, statistical steps employed for model cross-validation were outlined in the cross-validation script.